refactor: remove useless re-export patterns
