+++ src/screens/image-ocr/image-ocr-result.tsx
@@ -4,7 +4,7 @@ import { createStyles } from "antd-style";
 import { useResponsive } from "~/hooks/use-responsive";
 
 import { ImageOcrProcessing } from "./image-ocr-processing";
-import type { OcrResult, WorkerStatus } from "./image-ocr.utils";
+import type { OcrResult, WorkerStatus } from "./image-ocr.types";
 import { formatProcessingTime } from "./image-ocr.utils";
 
 const { Text, Paragraph } = Typography;
